Logica voor het maken van een api call overhevelen naar PokemonCard
