+++ .eslintcache
@@ -1,1 +1,1 @@
-[{"/Users/reinoptland/Code/Novi/example/src/index.js":"1","/Users/reinoptland/Code/Novi/example/src/App.js":"2","/Users/reinoptland/Code/Novi/example/src/reportWebVitals.js":"3","/Users/reinoptland/Code/Novi/example/src/components/PokemonCard.js":"4"},{"size":500,"mtime":1607450055695,"results":"5","hashOfConfig":"6"},{"size":1061,"mtime":1607690740687,"results":"7","hashOfConfig":"6"},{"size":362,"mtime":1607450055696,"results":"8","hashOfConfig":"6"},{"size":170,"mtime":1607690914320,"results":"9","hashOfConfig":"6"},{"filePath":"10","messages":"11","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"12"},"pa5su5",{"filePath":"13","messages":"14","errorCount":0,"warningCount":1,"fixableErrorCount":0,"fixableWarningCount":0,"source":null},{"filePath":"15","messages":"16","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"12"},{"filePath":"17","messages":"18","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"/Users/reinoptland/Code/Novi/example/src/index.js",[],["19","20"],"/Users/reinoptland/Code/Novi/example/src/App.js",["21"],"/Users/reinoptland/Code/Novi/example/src/reportWebVitals.js",[],"/Users/reinoptland/Code/Novi/example/src/components/PokemonCard.js",[],{"ruleId":"22","replacedBy":"23"},{"ruleId":"24","replacedBy":"25"},{"ruleId":"26","severity":1,"message":"27","line":38,"column":7,"nodeType":"28","endLine":38,"endColumn":51},"no-native-reassign",["29"],"no-negated-in-lhs",["30"],"jsx-a11y/alt-text","img elements must have an alt prop, either with meaningful text, or an empty string for decorative images.","JSXOpeningElement","no-global-assign","no-unsafe-negation"]
+[{"/Users/reinoptland/Code/Novi/example/src/index.js":"1","/Users/reinoptland/Code/Novi/example/src/App.js":"2","/Users/reinoptland/Code/Novi/example/src/reportWebVitals.js":"3","/Users/reinoptland/Code/Novi/example/src/components/PokemonCard.js":"4"},{"size":500,"mtime":1607450055695,"results":"5","hashOfConfig":"6"},{"size":276,"mtime":1607691208915,"results":"7","hashOfConfig":"6"},{"size":362,"mtime":1607450055696,"results":"8","hashOfConfig":"6"},{"size":987,"mtime":1607691171026,"results":"9","hashOfConfig":"6"},{"filePath":"10","messages":"11","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"12"},"pa5su5",{"filePath":"13","messages":"14","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"15","messages":"16","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"12"},{"filePath":"17","messages":"18","errorCount":0,"warningCount":1,"fixableErrorCount":0,"fixableWarningCount":0,"source":null},"/Users/reinoptland/Code/Novi/example/src/index.js",[],["19","20"],"/Users/reinoptland/Code/Novi/example/src/App.js",[],"/Users/reinoptland/Code/Novi/example/src/reportWebVitals.js",[],"/Users/reinoptland/Code/Novi/example/src/components/PokemonCard.js",["21"],{"ruleId":"22","replacedBy":"23"},{"ruleId":"24","replacedBy":"25"},{"ruleId":"26","severity":1,"message":"27","line":35,"column":7,"nodeType":"28","endLine":35,"endColumn":51},"no-native-reassign",["29"],"no-negated-in-lhs",["30"],"jsx-a11y/alt-text","img elements must have an alt prop, either with meaningful text, or an empty string for decorative images.","JSXOpeningElement","no-global-assign","no-unsafe-negation"]
